fix(apps): stop truncating longVersionCode to Int
`mapToAppInfo` read the platform's version code as
`packInfo.longVersionCode.toInt()`. `longVersionCode` is
`(versionCodeMajor shl 32) or versionCode`, so the cast discarded
versionCodeMajor outright and reinterpreted the low 32 bits as signed. An app
declaring versionCodeMajor = 1 with versionCode = 0 became 0; a code above
Int.MAX_VALUE wrapped negative.

That wrong number then propagated everywhere AppInfo.versionCode is used:

  - AppSorting sorted VERSION_CODE on the truncated value, so versionCodeMajor
    apps sorted below every ordinary app instead of above;
  - AppInfoDetailsScreen showed it to the user as the app's version code;
  - ApksMetadataGenerator wrote it into the `version_code` of the .apks/XAPK
    metadata Thor generates when exporting an app — which other installers, and
    Thor's own bundle analyzer, read back to decide update vs. downgrade;
  - the Room cache persisted it.

Widened to Long end to end: AppInfo.versionCode, AppEntity.versionCode and
ApksMetadata.versionCode. AppSorting and the details screen need no change (a
Comparator and a string template).

No Room migration or schema version bump: Room maps both Kotlin Int and Long to
SQLite affinity INTEGER, so the exported schema is byte-identical and the
identityHash is unchanged — verified by re-running KSP with --rerun-tasks and
diffing app/schemas (empty). Rows already cached with a truncated value keep it
until that package's lastUpdateTime changes and AppRepositoryImpl re-maps it;
that self-heals on the app's next update and only ever affected apps declaring
versionCodeMajor or a code above Int.MAX, so forcing a full re-scan on every
user would cost more than it fixes.

Also drops the `@Suppress("DEPRECATION")` on the AppInfo construction: it was a
leftover from the deprecated Int `packInfo.versionCode` this mapper used to
read, nothing in the expression is deprecated any more (verified by compiling
without it), and leaving it would swallow the next real deprecation.

Tests: AppSortingTest gains versionCodeMajor and above-Int.MAX ordering cases
(operands chosen so a truncating implementation inverts the result), and a new
ApksMetadataGeneratorTest pins the generated wire format. 83 unit tests pass;
lintFossDebug reports the same 5 pre-existing VectorPath issues; foss and store
release builds are green.

+++ app/src/main/java/com/valhalla/thor/data/source/local/room/AppEntity.kt
@@ -12,7 +12,14 @@ data class AppEntity(
     @PrimaryKey val packageName: String,
     val appName: String?,
     val versionName: String?,
-    val versionCode: Int,
+    // Int -> Long. No schema version bump or Migration is needed: Room maps both Kotlin Int and
+    // Long to SQLite affinity INTEGER, so the exported schema and its identityHash are unchanged
+    // (see app/schemas/.../5.json). Rows already cached with a truncated value keep it until the
+    // package's lastUpdateTime changes and AppRepositoryImpl re-maps it — acceptable, because the
+    // only values that were ever wrong belong to apps declaring versionCodeMajor or a code above
+    // Int.MAX, which are vanishingly rare, and forcing a full re-scan on every user to correct
+    // them would cost far more than it fixes.
+    val versionCode: Long,
     val minSdk: Int,
     val targetSdk: Int,
     val isSystem: Boolean,
